server/requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.40.0
